renamed all instances of GuageProgress component to Guage including tests
