+++ packages/core/src/components/ProgressBars/GaugeCard.tsx
@@ -18,7 +18,7 @@ import React, { FC } from 'react';
 import { makeStyles } from '@material-ui/core';
 import { InfoCard } from '../../layout/InfoCard';
 import { BottomLinkProps } from '../../layout/BottomLink';
-import { GaugeProgress } from './GaugeProgress';
+import { Gauge } from './Gauge';
 
 type Props = {
   title: string;
@@ -48,7 +48,7 @@ export const GaugeCard: FC<Props> = props => {
         deepLink={deepLink}
         variant={variant}
       >
-        <GaugeProgress value={progress} />
+        <Gauge value={progress} />
       </InfoCard>
     </div>
   );
